extract workrecord
